requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-stubs==5.0.0
django-stubs-ext==5.0.0
djangorestframework==3.15.1
djangorestframework-stubs==3.15.0
drf-nested-routers==0.93.5
idna==3.7
requests==2.31.0
sqlparse==0.5.0
tomli==2.0.1
types-PyYAML==6.0.12.20240311
types-requests==2.31.0.20240406
typing_extensions==4.11.0
urllib3==2.2.1
